Move files into src folder, add error middleware

+++ src/controllers/user.controller.js
@@ -4,16 +4,18 @@ import { userService } from "../services";
 import ApiError from "../utils/ApiError";
 import catchAsync from "../utils/CatchAsync";
 
-const getUser = async (req, res) => {
+const getUser = catchAsync(async (req, res) => {
   const user = await userService.getUserById(req.params.userId);
   if (!user) throw new ApiError(httpStatus.NOT_FOUND, "User can not found");
   res.send(user);
-};
+});
 
-const deleteUser = async (req, res) => {
+const deleteUser = catchAsync(async (req, res) => {
+  const user = await userService.getUserById(req.params.userId);
+  if (!user) throw new ApiError(httpStatus.NOT_FOUND, "User can not found");
   await userService.deleteUserById(req.params.userId);
   res.status(httpStatus.NO_CONTENT).send();
-};
+});
 
 module.exports = {
   getUser,
